backend/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/requirements.in
#
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.10
    # via -r requirements/requirements.in
aiosignal==1.3.2
    # via aiohttp
alembic==1.14.0
    # via -r requirements/requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
arxiv==2.1.3
    # via -r requirements/requirements.in
asyncio-throttle==1.0.2
    # via -r requirements/requirements.in
attrs==24.3.0
    # via aiohttp
biopython==1.84
    # via -r requirements/requirements.in
black==24.10.0
    # via -r requirements/requirements.in
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   pinecone-client
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   black
    #   uvicorn
fastapi==0.115.6
    # via -r requirements/requirements.in
feedparser==6.0.11
    # via arxiv
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flake8==7.1.1
    # via -r requirements/requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.10.0
    # via
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements/requirements.in
    #   semanticscholar
huggingface-hub==0.26.5
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via -r requirements/requirements.in
jinja2==3.1.4
    # via torch
joblib==1.4.2
    # via scikit-learn
mako==1.3.8
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mccabe==0.7.0
    # via flake8
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy==1.13.0
    # via -r requirements/requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nest-asyncio==1.6.0
    # via semanticscholar
networkx==3.4.2
    # via torch
numpy==2.2.0
    # via
    #   -r requirements/requirements.in
    #   biopython
    #   pandas
    #   scikit-learn
    #   scipy
    #   transformers
packaging==24.2
    # via
    #   black
    #   huggingface-hub
    #   pytest
    #   transformers
pandas==2.2.3
    # via -r requirements/requirements.in
pathspec==0.12.1
    # via black
pillow==11.0.0
    # via sentence-transformers
pinecone-client==5.0.1
    # via -r requirements/requirements.in
pinecone-plugin-inference==1.1.0
    # via pinecone-client
pinecone-plugin-interface==0.0.7
    # via
    #   pinecone-client
    #   pinecone-plugin-inference
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
psycopg2-binary==2.9.10
    # via -r requirements/requirements.in
pycodestyle==2.12.1
    # via flake8
pydantic==2.10.3
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.27.1
    # via pydantic
pydantic-settings==2.7.0
    # via -r requirements/requirements.in
pyflakes==3.2.0
    # via flake8
pytest==8.3.4
    # via
    #   -r requirements/requirements.in
    #   pytest-asyncio
pytest-asyncio==0.25.0
    # via -r requirements/requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   -r requirements/requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.19
    # via -r requirements/requirements.in
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements/requirements.in
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   arxiv
    #   huggingface-hub
    #   transformers
safetensors==0.4.5
    # via transformers
scikit-learn==1.6.0
    # via
    #   -r requirements/requirements.in
    #   sentence-transformers
scipy==1.14.1
    # via
    #   scikit-learn
    #   sentence-transformers
semanticscholar==0.9.0
    # via -r requirements/requirements.in
sentence-transformers==3.3.1
    # via -r requirements/requirements.in
sgmllib3k==1.0.0
    # via feedparser
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via
    #   -r requirements/requirements.in
    #   alembic
starlette==0.41.3
    # via fastapi
sympy==1.13.1
    # via torch
tenacity==9.0.0
    # via
    #   -r requirements/requirements.in
    #   semanticscholar
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.21.0
    # via transformers
torch==2.5.1
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   pinecone-client
    #   sentence-transformers
    #   transformers
transformers==4.47.0
    # via sentence-transformers
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   huggingface-hub
    #   mypy
    #   pinecone-client
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via
    #   pinecone-client
    #   requests
uvicorn[standard]==0.34.0
    # via -r requirements/requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.3
    # via uvicorn
websockets==14.1
    # via uvicorn
yarl==1.18.3
    # via aiohttp
